Reduce over-modularization
