--- core/package.md
@@ -1,3 +1,0 @@
-# Package com.juul.stropping
-
-`implementation "com.github.juullabs-oss.stropping:core:$version"`
