+++ stropping/.gitignore
