+++ stropping/proguard-rules.pro
