+++ stropping/src/androidTest/java/com/juul/stropping/kodein/ExampleInstrumentedTest.kt
