+++ stropping/src/main/AndroidManifest.xml
